fuel_calc/backend/requirements.txt
Django==4.2
dj-database-url==2.1.0
psycopg2-binary
djangorestframework
drf_yasg
django-cors-headers
Pillow
